refactor rename deployment.ts -> deploymentUpdater.ts
