+++ src/background/background.ts
@@ -36,7 +36,7 @@ import initNavigation from "@/background/navigation";
 import initGoogle, { isGAPISupported } from "@/contrib/google/initGoogle";
 import initExecutor from "@/background/executor";
 import initBrowserCommands from "@/background/initBrowserCommands";
-import initDeploymentUpdater from "@/background/deployment";
+import initDeploymentUpdater from "@/background/deploymentUpdater";
 import initFirefoxCompat from "@/background/firefoxCompat";
 import activateBrowserActionIcon from "@/background/activateBrowserActionIcon";
 import initActiveTabTracking from "@/background/activeTab";
